coser_bot/requirements.txt
# 核心依赖
fastapi==0.109.2
uvicorn==0.27.1
python-telegram-bot==20.8
pydantic-settings==2.2.1
sqlalchemy==2.0.27
aiomysql==0.2.0
redis==5.0.1
motor==3.3.2
aio-pika==9.4.0
aiosmtplib==3.0.1

# 测试依赖
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
locust==2.24.0
